fix: migrate to Bazel 9 (removed rules, sandboxed test strategy)
Bazel 9 removed several native rule families from the default BUILD
namespace. Migrate to the stubbed cc_rules module path and avoid the
sandbox-inside-sandbox issue when running tests under Zed's terminal
sandbox by using the standalone TestRunner strategy.

- .bazelrc: pin test strategy to standalone to allow tests to run in
  nested sandbox environments (e.g. Zed terminal sandbox).
- MODULE.bazel: drop bazel_dep(cpp-httplib) in favor of http_archive
  with a patched BUILD content that declares cc_library via @rules_cc.
  Add rules_cc to bazel_dep.
- MODULE.bazel & BUILD.bazel: load cc_library/cc_binary/cc_test from
  @rules_cc//cc:defs.bzl; adjust httplib label to @cpp_httplib.
- MODULE.bazel: patch dotenv_cpp injected BUILD to use rules_cc.
- MODULE.bazel.lock: updated for the new dependency graph.
- BUILD.bazel: add -lssl/-lcrypto linkopts to the aoc library (needed
  now that cpp-httplib is no longer dragging boringssl in bzlmod mode).
- README.md: use bash directly for the aoc-init scaffolding helper,
  since the sh_binary wrapper was removed.
- include/solutions/day01.hpp: widen PASSWORD_METHOD to long to silence
  overflow truncation warning when comparing with the 5-byte CLICK sentinel.

+++ 2025/include/solutions/day01.hpp
@@ -9,7 +9,7 @@
 
 namespace day01 {
 
-  using PASSWORD_METHOD = int;
+  using PASSWORD_METHOD = long;
 
   const PASSWORD_METHOD d1p1 = 0x0;
   const PASSWORD_METHOD d1p2 = 0x434C49434B;
